requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
face-recognition>=1.3.0
deep-sort-realtime>=1.3.2
Pillow>=10.0.0
ultralytics>=8.0.0 